tipos alterados para double, tratando o erro de divisao por zero

+++ Calculadora/Operacoes.cs
@@ -3,10 +3,10 @@ namespace Calculadora
 {
     public class Operacoes
     {
-        public long valorA {get;set;}
+        public double valorA {get;set;}
         public char operador { get; set; }
-        public long valorB { get; set; }
-        public long resultado { get; set; }
+        public double valorB { get; set; }
+        public double resultado { get; set; }
 
     }
 }
